src/requirements.txt
ipython==8.12.3
matplotlib==3.10.0
nodegam==0.3.0
numpy==2.2.3
pandas==2.2.3
pyro_ppl==1.9.1
qhoptim==1.1.0
ray==2.42.1
Requests==2.32.3
scikit_learn==1.6.1
torch==2.6.0
tqdm==4.67.1
xgboost==2.1.4
